ideation_agent/requirements.txt
# Core dependencies
fastapi>=0.115.12,<0.116.0
uvicorn[standard]>=0.18.0,<0.21.0
sqlmodel>=0.0.8,<0.1.0

# Use the psycopg sync driver for Postgres
psycopg[binary]>=3.1.0,<4.0.0

# Environment management and OpenAI
python-dotenv>=1.0.0,<2.0.0
openai>=1.79.0,<2.0.0
pydantic>=1.10.12,<2.0.0

# HTTP & ASGI support
httpx>=0.28.1,<0.29.0
anyio>=4.9.0,<5.0.0
sniffio>=1.3.1
h11>=0.16.0

# Utilities
certifi>=2025.4.26
click>=8.2.0
tqdm>=4.67.1
typing-inspection>=0.4.0
typing-extensions>=4.13.2
greenlet>=2.0.0